lexaro-api/neural-ocr/requirements.txt
fastapi
uvicorn
easyocr
pillow
numpy
opencv-python-headless
